add test_query_live_cells.
