+++ core/storage/src/relational/tests/mod.rs
@@ -1,4 +1,4 @@
-mod fetch_distinct_tx_hashes_count_test;
+mod fetch_mod_test;
 mod get_block_test;
 mod get_cell_test;
 mod get_historical_live_cells_test;
